+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Fixed a bug where `Craft.BaseElementIndexView::this.canSelectElement()` wasn’t getting applied for lazy-loaded elements.
 - Fixed a bug where element thumbnails weren’t getting loaded for expanded relational field previews within element indexes.
 - Fixed an error that occurred when deleting a volume with a missing filesystem type.